Revert "Show ANT+ manufacturer and model name"
This reverts commit d90e7023183a99fb07e08df034c07ce98499dceb.
Upstream j-antplus doesn't have the changes for this.
I hope I can find my branch with them.
