+++ src/me/danielschaefer/sensorwrangler/sensors/AntHeartRateSensor.kt
@@ -6,8 +6,6 @@ import be.glever.ant.usb.AntUsbDevice
 import be.glever.antplus.common.datapage.AbstractAntPlusDataPage
 import be.glever.antplus.hrm.HRMChannel
 import be.glever.antplus.hrm.datapage.HrmDataPageRegistry
-import be.glever.antplus.hrm.datapage.background.HrmDataPage2ManufacturerInformation
-import be.glever.antplus.hrm.datapage.background.HrmDataPage3ProductInformation
 import be.glever.antplus.hrm.datapage.main.HrmDataPage4PreviousHeartBeatEvent
 import javafx.application.Platform
 import me.danielschaefer.sensorwrangler.Measurement
@@ -23,29 +21,18 @@ class AntHeartRateSensor : AntPlusSensor<HRMChannel>() {
     }
     override val measurements: List<Measurement> = listOf(measurement)
 
-    override fun handleDevSpecificMessage(antMessage: AntMessage?) {
+    override fun handleMessage(antMessage: AntMessage?) {
         if (antMessage is BroadcastDataMessage) {
             val payLoad = antMessage.payLoad
             removeToggleBit(payLoad)
             val dataPage: AbstractAntPlusDataPage = registry.constructDataPage(payLoad)
 
-            when (dataPage) {
-                is HrmDataPage2ManufacturerInformation ->
-                    if (manufacturerIdProperty.value == null) {
-                        manufacturerIdProperty.value = dataPage.manufacturerId
-                        // Assign it again, to re-fire any listener. Because:
-                        // The model name can only be determined once the manufacturer ID has been determined
-                        modelNumberProperty.value = modelNumberProperty.value
-                    }
-                is HrmDataPage3ProductInformation ->
-                    if (modelNumberProperty.value == null)
-                        modelNumberProperty.value = dataPage.modelNumber.toInt()
-                is HrmDataPage4PreviousHeartBeatEvent ->
-                    // TODO: Call runLater in UI code, not here
-                    // UI can only be updated from UI threads
-                    Platform.runLater {
-                        measurement.addDataPoint(dataPage.computedHeartRateInBpm.toDouble())
-                    }
+            if (dataPage is HrmDataPage4PreviousHeartBeatEvent) {
+                // TODO: Call runLater in UI code, not here
+                // UI can only be updated from UI threads
+                Platform.runLater {
+                    measurement.addDataPoint(dataPage.computedHeartRateInBpm.toDouble())
+                }
             }
         }
     }
